adding convolution and spectral reorg support and continue to move functions out of signaltools
