+++ python/cusignal/convolution/convolution_utils.py
@@ -1,0 +1,214 @@
+# Copyright (c) 2019-2020, NVIDIA CORPORATION.
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from cupy import ndarray, array, asarray
+import cupy as cp
+
+import timeit
+
+
+_modedict = {"valid": 0, "same": 1, "full": 2}
+
+_boundarydict = {
+    "fill": 0,
+    "pad": 0,
+    "wrap": 2,
+    "circular": 2,
+    "symm": 1,
+    "symmetric": 1,
+    "reflect": 4,
+}
+
+
+def _inputs_swap_needed(mode, shape1, shape2):
+    """
+    If in 'valid' mode, returns whether or not the input arrays need to be
+    swapped depending on whether `shape1` is at least as large as `shape2` in
+    every dimension.
+
+    This is important for some of the correlation and convolution
+    implementations in this module, where the larger array input needs to come
+    before the smaller array input when operating in this mode.
+
+    Note that if the mode provided is not 'valid', False is immediately
+    returned.
+    """
+    if mode == "valid":
+        ok1, ok2 = True, True
+
+        for d1, d2 in zip(shape1, shape2):
+            if not d1 >= d2:
+                ok1 = False
+            if not d2 >= d1:
+                ok2 = False
+
+        if not (ok1 or ok2):
+            raise ValueError(
+                "For 'valid' mode, one must be at least "
+                "as large as the other in every dimension"
+            )
+
+        return not ok1
+
+    return False
+
+
+def _numeric_arrays(arrays, kinds="buifc"):
+    """
+    See if a list of arrays are all numeric.
+
+    Parameters
+    ----------
+    ndarrays : array or list of arrays
+        arrays to check if numeric.
+    numeric_kinds : string-like
+        The dtypes of the arrays to be checked. If the dtype.kind of
+        the ndarrays are not in this string the function returns False and
+        otherwise returns True.
+    """
+    if type(arrays) == ndarray:
+        return arrays.dtype.kind in kinds
+    for array_ in arrays:
+        if array_.dtype.kind not in kinds:
+            return False
+    return True
+
+
+def _centered(arr, newshape):
+    # Return the center newshape portion of the array.
+    newshape = asarray(newshape)
+    currshape = array(arr.shape)
+    startind = (currshape - newshape) // 2
+    endind = startind + newshape
+    myslice = [slice(startind[k], endind[k]) for k in range(len(endind))]
+    return arr[tuple(myslice)]
+
+
+def _prod(iterable):
+    """
+    Product of a list of numbers.
+    Faster than np.prod for short lists like array shapes.
+    """
+    product = 1
+    for x in iterable:
+        product *= x
+    return product
+
+
+def _fftconv_faster(x, h, mode):
+    """
+    See if using `fftconvolve` or `_correlateND` is faster. The boolean value
+    returned depends on the sizes and shapes of the input values.
+
+    The big O ratios were found to hold across different machines, which makes
+    sense as it's the ratio that matters (the effective speed of the computer
+    is found in both big O constants). Regardless, this had been tuned on an
+    early 2015 MacBook Pro with 8GB RAM and an Intel i5 processor.
+    """
+    if mode == "full":
+        out_shape = [n + k - 1 for n, k in zip(x.shape, h.shape)]
+        big_O_constant = 10963.92823819 if x.ndim == 1 else 8899.1104874
+    elif mode == "same":
+        out_shape = x.shape
+        if x.ndim == 1:
+            if h.size <= x.size:
+                big_O_constant = 7183.41306773
+            else:
+                big_O_constant = 856.78174111
+        else:
+            big_O_constant = 34519.21021589
+    elif mode == "valid":
+        out_shape = [n - k + 1 for n, k in zip(x.shape, h.shape)]
+        big_O_constant = 41954.28006344 if x.ndim == 1 else 66453.24316434
+    else:
+        raise ValueError(
+            "Acceptable mode flags are \
+                         'valid',"
+            " 'same', or 'full'."
+        )
+
+    # see whether the Fourier transform convolution method or the direct
+    # convolution method is faster (discussed in scikit-image PR #1792)
+    direct_time = x.size * h.size * _prod(out_shape)
+    fft_time = sum(
+        n * cp.log(n) for n in (x.shape + h.shape + tuple(out_shape))
+    )
+
+    return big_O_constant * fft_time < direct_time
+
+
+def _timeit_fast(stmt="pass", setup="pass", repeat=3):
+    """
+    Returns the time the statement/function took, in seconds.
+
+    Faster, less precise version of IPython's timeit. `stmt` can be a statement
+    written as a string or a callable.
+
+    Will do only 1 loop (like IPython's timeit) with no repetitions
+    (unlike IPython) for very slow functions.  For fast functions, only does
+    enough loops to take 5 ms, which seems to produce similar results (on
+    Windows at least), and avoids doing an extraneous cycle that isn't
+    measured.
+
+    """
+    timer = timeit.Timer(stmt, setup)
+
+    # determine number of calls per rep so total time for 1 rep >= 5 ms
+    x = 0
+    for p in range(0, 10):
+        number = 10 ** p
+        x = timer.timeit(number)  # seconds
+        if x >= 5e-3 / 10:  # 5 ms for final test, 1/10th that for this one
+            break
+    if x > 1:  # second
+        # If it's macroscopic, don't bother with repetitions
+        best = x
+    else:
+        number *= 10
+        r = timer.repeat(repeat, number)
+        best = min(r)
+
+    sec = best / number
+    return sec
+
+
+def _valfrommode(mode):
+    try:
+        return _modedict[mode]
+    except KeyError:
+        raise ValueError(
+            "Acceptable mode flags are 'valid'," " 'same', or 'full'."
+        )
+
+
+def _bvalfromboundary(boundary):
+    try:
+        return _boundarydict[boundary] << 2
+    except KeyError:
+        raise ValueError(
+            "Acceptable boundary flags are 'fill', 'circular' "
+            "(or 'wrap'), and 'symmetric' (or 'symm')."
+        )
+
+
+def _iDivUp(a, b):
+    return (a // b + 1) if (a % b != 0) else (a // b)
+
+
+def _reverse_and_conj(x):
+    """
+    Reverse array `x` in all dimensions and perform the complex conjugate
+    """
+    reverse = (slice(None, None, -1),) * x.ndim
+    # return cp.flip(x, 0)
+    return x[reverse].conj()
