+++ python/cusignal/signaltools.py
@@ -12,1263 +12,6 @@
 # limitations under the License.
 
 import cupy as cp
-from cupy import (
-    angle,
-    arange,
-    argsort,
-    array,
-    asarray,
-    atleast_2d,
-    dot,
-    exp,
-    expand_dims,
-    iscomplexobj,
-    mean,
-    ndarray,
-    newaxis,
-    ones,
-    pi,
-    prod,
-    r_,
-    ravel,
-    reshape,
-    sort,
-    take,
-    transpose,
-    unique,
-    where,
-    zeros,
-)
-from cupyx.scipy import fftpack
-from cupy.fft import ifftshift
-from cupy import linalg
-import numpy as np
-
-from math import gcd
-import timeit
-import sys
-
-from . import _signaltools
-from .window_functions.windows import get_window
-from .fftpack_helper import _init_nd_shape_and_axes_sorted, next_fast_len
-from ._upfirdn import upfirdn, _output_len
-from .fir_filter_design import firwin
-
-_modedict = {"valid": 0, "same": 1, "full": 2}
-
-
-def _inputs_swap_needed(mode, shape1, shape2):
-    """
-    If in 'valid' mode, returns whether or not the input arrays need to be
-    swapped depending on whether `shape1` is at least as large as `shape2` in
-    every dimension.
-
-    This is important for some of the correlation and convolution
-    implementations in this module, where the larger array input needs to come
-    before the smaller array input when operating in this mode.
-
-    Note that if the mode provided is not 'valid', False is immediately
-    returned.
-    """
-    if mode == "valid":
-        ok1, ok2 = True, True
-
-        for d1, d2 in zip(shape1, shape2):
-            if not d1 >= d2:
-                ok1 = False
-            if not d2 >= d1:
-                ok2 = False
-
-        if not (ok1 or ok2):
-            raise ValueError(
-                "For 'valid' mode, one must be at least "
-                "as large as the other in every dimension"
-            )
-
-        return not ok1
-
-    return False
-
-
-def correlate(in1, in2, mode="full", method="auto"):
-    r"""
-    Cross-correlate two N-dimensional arrays.
-
-    Cross-correlate `in1` and `in2`, with the output size determined by the
-    `mode` argument.
-
-    Parameters
-    ----------
-    in1 : array_like
-        First input.
-    in2 : array_like
-        Second input. Should have the same number of dimensions as `in1`.
-    mode : str {'full', 'valid', 'same'}, optional
-        A string indicating the size of the output:
-
-        ``full``
-           The output is the full discrete linear cross-correlation
-           of the inputs. (Default)
-        ``valid``
-           The output consists only of those elements that do not
-           rely on the zero-padding. In 'valid' mode, either `in1` or `in2`
-           must be at least as large as the other in every dimension.
-        ``same``
-           The output is the same size as `in1`, centered
-           with respect to the 'full' output.
-    method : str {'auto', 'direct', 'fft'}, optional
-        A string indicating which method to use to calculate the correlation.
-
-        ``direct``
-           The correlation is determined directly from sums, the definition of
-           correlation.
-        ``fft``
-           The Fast Fourier Transform is used to perform the correlation more
-           quickly (only available for numerical arrays.)
-        ``auto``
-           Automatically chooses direct or Fourier method based on an estimate
-           of which is faster (default).  See `convolve` Notes for more detail.
-
-           .. versionadded:: 0.19.0
-
-    Returns
-    -------
-    correlate : array
-        An N-dimensional array containing a subset of the discrete linear
-        cross-correlation of `in1` with `in2`.
-
-    See Also
-    --------
-    choose_conv_method : contains more documentation on `method`.
-
-    Notes
-    -----
-    The correlation z of two d-dimensional arrays x and y is defined as::
-
-        z[...,k,...] =
-            sum[..., i_l, ...] x[..., i_l,...] * conj(y[..., i_l - k,...])
-
-    This way, if x and y are 1-D arrays and ``z = correlate(x, y, 'full')``
-    then
-
-    .. math::
-
-          z[k] = (x * y)(k - N + 1)
-               = \sum_{l=0}^{||x||-1}x_l y_{l-k+N-1}^{*}
-
-    for :math:`k = 0, 1, ..., ||x|| + ||y|| - 2`
-
-    where :math:`||x||` is the length of ``x``, :math:`N = \max(||x||,||y||)`,
-    and :math:`y_m` is 0 when m is outside the range of y.
-
-    ``method='fft'`` only works for numerical arrays as it relies on
-    `fftconvolve`. In certain cases (i.e., arrays of objects or when
-    rounding integers can lose precision), ``method='direct'`` is always used.
-
-    Examples
-    --------
-    Implement a matched filter using cross-correlation, to recover a signal
-    that has passed through a noisy channel.
-
-    >>> from scipy import signal
-    >>> sig = np.repeat([0., 1., 1., 0., 1., 0., 0., 1.], 128)
-    >>> sig_noise = sig + np.random.randn(len(sig))
-    >>> corr = signal.correlate(sig_noise, np.ones(128), mode='same') / 128
-
-    >>> import matplotlib.pyplot as plt
-    >>> clock = np.arange(64, len(sig), 128)
-    >>> fig, (ax_orig, ax_noise, ax_corr) = plt.subplots(3, 1, sharex=True)
-    >>> ax_orig.plot(sig)
-    >>> ax_orig.plot(clock, sig[clock], 'ro')
-    >>> ax_orig.set_title('Original signal')
-    >>> ax_noise.plot(sig_noise)
-    >>> ax_noise.set_title('Signal with noise')
-    >>> ax_corr.plot(corr)
-    >>> ax_corr.plot(clock, corr[clock], 'ro')
-    >>> ax_corr.axhline(0.5, ls=':')
-    >>> ax_corr.set_title('Cross-correlated with rectangular pulse')
-    >>> ax_orig.margins(0, 0.1)
-    >>> fig.tight_layout()
-    >>> fig.show()
-
-    """
-    in1 = asarray(in1)
-    in2 = asarray(in2)
-
-    if in1.ndim == in2.ndim == 0:
-        return in1 * in2.conj()
-    elif in1.ndim != in2.ndim:
-        raise ValueError("in1 and in2 should have the same dimensionality")
-
-    # this either calls fftconvolve or this function with method=='direct'
-    if method in ("fft", "auto"):
-        return convolve(in1, _reverse_and_conj(in2), mode, method)
-
-    elif method == "direct":
-
-        if in1.ndim > 1:
-            raise ValueError("Direct method is only implemented for 1D")
-
-        swapped_inputs = in2.size > in1.size
-
-        if swapped_inputs:
-            in1, in2 = in2, in1
-
-        return _signaltools._convolve(in1, in2, False, swapped_inputs, mode)
-
-    else:
-        raise ValueError(
-            "Acceptable method flags are 'auto'," " 'direct', or 'fft'."
-        )
-
-
-def _centered(arr, newshape):
-    # Return the center newshape portion of the array.
-    newshape = asarray(newshape)
-    currshape = array(arr.shape)
-    startind = (currshape - newshape) // 2
-    endind = startind + newshape
-    myslice = [slice(startind[k], endind[k]) for k in range(len(endind))]
-    return arr[tuple(myslice)]
-
-
-def fftconvolve(in1, in2, mode="full", axes=None):
-    """Convolve two N-dimensional arrays using FFT.
-
-    Convolve `in1` and `in2` using the fast Fourier transform method, with
-    the output size determined by the `mode` argument.
-
-    This is generally much faster than `convolve` for large arrays (n > ~500),
-    but can be slower when only a few output values are needed, and can only
-    output float arrays (int or object array inputs will be cast to float).
-
-    As of v0.19, `convolve` automatically chooses this method or the direct
-    method based on an estimation of which is faster.
-
-    Parameters
-    ----------
-    in1 : array_like
-        First input.
-    in2 : array_like
-        Second input. Should have the same number of dimensions as `in1`.
-    mode : str {'full', 'valid', 'same'}, optional
-        A string indicating the size of the output:
-
-        ``full``
-           The output is the full discrete linear convolution
-           of the inputs. (Default)
-        ``valid``
-           The output consists only of those elements that do not
-           rely on the zero-padding. In 'valid' mode, either `in1` or `in2`
-           must be at least as large as the other in every dimension.
-        ``same``
-           The output is the same size as `in1`, centered
-           with respect to the 'full' output.
-           axis : tuple, optional
-    axes : int or array_like of ints or None, optional
-        Axes over which to compute the convolution.
-        The default is over all axes.
-
-
-    Returns
-    -------
-    out : array
-        An N-dimensional array containing a subset of the discrete linear
-        convolution of `in1` with `in2`.
-
-    Examples
-    --------
-    Autocorrelation of white noise is an impulse.
-
-    >>> from scipy import signal
-    >>> sig = np.random.randn(1000)
-    >>> autocorr = signal.fftconvolve(sig, sig[::-1], mode='full')
-
-    >>> import matplotlib.pyplot as plt
-    >>> fig, (ax_orig, ax_mag) = plt.subplots(2, 1)
-    >>> ax_orig.plot(sig)
-    >>> ax_orig.set_title('White noise')
-    >>> ax_mag.plot(np.arange(-len(sig)+1,len(sig)), autocorr)
-    >>> ax_mag.set_title('Autocorrelation')
-    >>> fig.tight_layout()
-    >>> fig.show()
-
-    Gaussian blur implemented using FFT convolution.  Notice the dark borders
-    around the image, due to the zero-padding beyond its boundaries.
-    The `convolve2d` function allows for other types of image boundaries,
-    but is far slower.
-
-    >>> from scipy import misc
-    >>> face = misc.face(gray=True)
-    >>> kernel = np.outer(signal.gaussian(70, 8), signal.gaussian(70, 8))
-    >>> blurred = signal.fftconvolve(face, kernel, mode='same')
-
-    >>> fig, (ax_orig, ax_kernel, ax_blurred) = plt.subplots(3, 1,
-    ...                                                      figsize=(6, 15))
-    >>> ax_orig.imshow(face, cmap='gray')
-    >>> ax_orig.set_title('Original')
-    >>> ax_orig.set_axis_off()
-    >>> ax_kernel.imshow(kernel, cmap='gray')
-    >>> ax_kernel.set_title('Gaussian kernel')
-    >>> ax_kernel.set_axis_off()
-    >>> ax_blurred.imshow(blurred, cmap='gray')
-    >>> ax_blurred.set_title('Blurred')
-    >>> ax_blurred.set_axis_off()
-    >>> fig.show()
-
-    """
-    in1 = asarray(in1)
-    in2 = asarray(in2)
-    noaxes = axes is None
-
-    if in1.ndim == in2.ndim == 0:  # scalar inputs
-        return in1 * in2
-    elif in1.ndim != in2.ndim:
-        raise ValueError("in1 and in2 should have the same dimensionality")
-    elif in1.size == 0 or in2.size == 0:  # empty arrays
-        return array([])
-
-    _, axes = _init_nd_shape_and_axes_sorted(in1, shape=None, axes=axes)
-    # axes needs to be numpy type for proper execution in FFT
-    axes = cp.asnumpy(axes)
-
-    if not noaxes and not axes.size:
-        raise ValueError("when provided, axes cannot be empty")
-
-    if noaxes:
-        other_axes = array([], dtype=cp.intc)
-    else:
-        other_axes = cp.setdiff1d(arange(in1.ndim), axes)
-
-    s1 = array(in1.shape)
-    s2 = array(in2.shape)
-
-    if not cp.all(
-        (s1[other_axes] == s2[other_axes])
-        | (s1[other_axes] == 1)
-        | (s2[other_axes] == 1)
-    ):
-        raise ValueError(
-            "incompatible shapes for in1 and in2:"
-            " {0} and {1}".format(in1.shape, in2.shape)
-        )
-
-    complex_result = cp.issubdtype(
-        in1.dtype, cp.complexfloating
-    ) or cp.issubdtype(in2.dtype, cp.complexfloating)
-    shape = cp.maximum(s1, s2)
-    shape[axes] = s1[axes] + s2[axes] - 1
-
-    # Check that input sizes are compatible with 'valid' mode
-    if _inputs_swap_needed(mode, s1, s2):
-        # Convolution is commutative; order doesn't have any effect on output
-        in1, s1, in2, s2 = in2, s2, in1, s1
-
-    # Speed up FFT by padding to optimal size for FFTPACK
-    fshape = [next_fast_len(d) for d in shape[axes]]
-    fslice = tuple([slice(sz) for sz in shape])
-
-    if not complex_result:
-        sp1 = cp.fft.rfftn(in1, fshape, axes=axes)
-        sp2 = cp.fft.rfftn(in2, fshape, axes=axes)
-        ret = cp.fft.irfftn(sp1 * sp2, fshape, axes=axes)[fslice].copy()
-    else:
-        # If we're here, it's either because we need a complex result, or we
-        # failed to acquire _rfft_lock (meaning rfftn isn't threadsafe and
-        # is already in use by another thread).  In either case, use the
-        # (threadsafe but slower) SciPy complex-FFT routines instead.
-        sp1 = fftpack.fftn(in1, fshape, axes=axes)
-        sp2 = fftpack.fftn(in2, fshape, axes=axes)
-        ret = fftpack.ifftn(sp1 * sp2, axes=axes)[fslice].copy()
-        if not complex_result:
-            ret = ret.real
-
-    if mode == "full":
-        return ret
-    elif mode == "same":
-        return _centered(ret, s1)
-    elif mode == "valid":
-        shape_valid = shape.copy()
-        shape_valid[axes] = s1[axes] - s2[axes] + 1
-        return _centered(ret, shape_valid)
-    else:
-        raise ValueError(
-            "acceptable mode flags are \
-                         'valid',"
-            " 'same', or 'full'"
-        )
-
-
-def _numeric_arrays(arrays, kinds="buifc"):
-    """
-    See if a list of arrays are all numeric.
-
-    Parameters
-    ----------
-    ndarrays : array or list of arrays
-        arrays to check if numeric.
-    numeric_kinds : string-like
-        The dtypes of the arrays to be checked. If the dtype.kind of
-        the ndarrays are not in this string the function returns False and
-        otherwise returns True.
-    """
-    if type(arrays) == ndarray:
-        return arrays.dtype.kind in kinds
-    for array_ in arrays:
-        if array_.dtype.kind not in kinds:
-            return False
-    return True
-
-
-def _fftconv_faster(x, h, mode):
-    """
-    See if using `fftconvolve` or `_correlateND` is faster. The boolean value
-    returned depends on the sizes and shapes of the input values.
-
-    The big O ratios were found to hold across different machines, which makes
-    sense as it's the ratio that matters (the effective speed of the computer
-    is found in both big O constants). Regardless, this had been tuned on an
-    early 2015 MacBook Pro with 8GB RAM and an Intel i5 processor.
-    """
-    if mode == "full":
-        out_shape = [n + k - 1 for n, k in zip(x.shape, h.shape)]
-        big_O_constant = 10963.92823819 if x.ndim == 1 else 8899.1104874
-    elif mode == "same":
-        out_shape = x.shape
-        if x.ndim == 1:
-            if h.size <= x.size:
-                big_O_constant = 7183.41306773
-            else:
-                big_O_constant = 856.78174111
-        else:
-            big_O_constant = 34519.21021589
-    elif mode == "valid":
-        out_shape = [n - k + 1 for n, k in zip(x.shape, h.shape)]
-        big_O_constant = 41954.28006344 if x.ndim == 1 else 66453.24316434
-    else:
-        raise ValueError(
-            "Acceptable mode flags are \
-                         'valid',"
-            " 'same', or 'full'."
-        )
-
-    # see whether the Fourier transform convolution method or the direct
-    # convolution method is faster (discussed in scikit-image PR #1792)
-    direct_time = x.size * h.size * _prod(out_shape)
-    fft_time = sum(
-        n * cp.log(n) for n in (x.shape + h.shape + tuple(out_shape))
-    )
-
-    return big_O_constant * fft_time < direct_time
-
-
-#  TODO: Does this even work????
-def _reverse_and_conj(x):
-    """
-    Reverse array `x` in all dimensions and perform the complex conjugate
-    """
-    reverse = (slice(None, None, -1),) * x.ndim
-    # return cp.flip(x, 0)
-    return x[reverse].conj()
-
-
-def _np_conv_ok(volume, kernel, mode):
-    """
-    See if numpy supports convolution of `volume` and `kernel` (i.e. both are
-    1D ndarrays and of the appropriate shape).  NumPy's 'same' mode uses the
-    size of the larger input, while SciPy's uses the size of the first input.
-
-    Invalid mode strings will return False and be caught by the calling func.
-    """
-    if volume.ndim == kernel.ndim == 1:
-        if mode in ("full", "valid"):
-            return True
-        elif mode == "same":
-            return volume.size >= kernel.size
-    else:
-        return False
-
-
-def _timeit_fast(stmt="pass", setup="pass", repeat=3):
-    """
-    Returns the time the statement/function took, in seconds.
-
-    Faster, less precise version of IPython's timeit. `stmt` can be a statement
-    written as a string or a callable.
-
-    Will do only 1 loop (like IPython's timeit) with no repetitions
-    (unlike IPython) for very slow functions.  For fast functions, only does
-    enough loops to take 5 ms, which seems to produce similar results (on
-    Windows at least), and avoids doing an extraneous cycle that isn't
-    measured.
-
-    """
-    timer = timeit.Timer(stmt, setup)
-
-    # determine number of calls per rep so total time for 1 rep >= 5 ms
-    x = 0
-    for p in range(0, 10):
-        number = 10 ** p
-        x = timer.timeit(number)  # seconds
-        if x >= 5e-3 / 10:  # 5 ms for final test, 1/10th that for this one
-            break
-    if x > 1:  # second
-        # If it's macroscopic, don't bother with repetitions
-        best = x
-    else:
-        number *= 10
-        r = timer.repeat(repeat, number)
-        best = min(r)
-
-    sec = best / number
-    return sec
-
-
-def choose_conv_method(in1, in2, mode="full", measure=False):
-    """
-    Find the fastest convolution/correlation method.
-
-    This primarily exists to be called during the ``method='auto'`` option in
-    `convolve` and `correlate`, but can also be used when performing many
-    convolutions of the same input shapes and dtypes, determining
-    which method to use for all of them, either to avoid the overhead of the
-    'auto' option or to use accurate real-world measurements.
-
-    Parameters
-    ----------
-    in1 : array_like
-        The first argument passed into the convolution function.
-    in2 : array_like
-        The second argument passed into the convolution function.
-    mode : str {'full', 'valid', 'same'}, optional
-        A string indicating the size of the output:
-
-        ``full``
-           The output is the full discrete linear convolution
-           of the inputs. (Default)
-        ``valid``
-           The output consists only of those elements that do not
-           rely on the zero-padding.
-        ``same``
-           The output is the same size as `in1`, centered
-           with respect to the 'full' output.
-    measure : bool, optional
-        If True, run and time the convolution of `in1` and `in2` with both
-        methods and return the fastest. If False (default), predict the fastest
-        method using precomputed values.
-
-    Returns
-    -------
-    method : str
-        A string indicating which convolution method is fastest, either
-        'direct' or 'fft'
-    times : dict, optional
-        A dictionary containing the times (in seconds) needed for each method.
-        This value is only returned if ``measure=True``.
-
-    See Also
-    --------
-    convolve
-    correlate
-
-    Notes
-    -----
-    For large n, ``measure=False`` is accurate and can quickly determine the
-    fastest method to perform the convolution.  However, this is not as
-    accurate for small n (when any dimension in the input or output is small).
-
-    In practice, we found that this function estimates the faster method up to
-    a multiplicative factor of 5 (i.e., the estimated method is *at most* 5
-    times slower than the fastest method). The estimation values were tuned on
-    an early 2015 MacBook Pro with 8GB RAM but we found that the prediction
-    held *fairly* accurately across different machines.
-
-    If ``measure=True``, time the convolutions. Because this function uses
-    `fftconvolve`, an error will be thrown if it does not support the inputs.
-    There are cases when `fftconvolve` supports the inputs but this function
-    returns `direct` (e.g., to protect against floating point integer
-    precision).
-
-    .. versionadded:: 0.19
-
-    Examples
-    --------
-    Estimate the fastest method for a given input:
-
-    >>> from scipy import signal
-    >>> a = np.random.randn(1000)
-    >>> b = np.random.randn(1000000)
-    >>> method = signal.choose_conv_method(a, b, mode='same')
-    >>> method
-    'fft'
-
-    This can then be applied to other arrays of the same dtype and shape:
-
-    >>> c = np.random.randn(1000)
-    >>> d = np.random.randn(1000000)
-    >>> # `method` works with correlate and convolve
-    >>> corr1 = signal.correlate(a, b, mode='same', method=method)
-    >>> corr2 = signal.correlate(c, d, mode='same', method=method)
-    >>> conv1 = signal.convolve(a, b, mode='same', method=method)
-    >>> conv2 = signal.convolve(c, d, mode='same', method=method)
-
-    """
-    volume = asarray(in1)
-    kernel = asarray(in2)
-
-    if measure:
-        times = {}
-        for method in ["fft", "direct"]:
-            times[method] = _timeit_fast(
-                lambda: convolve(volume, kernel, mode=mode, method=method)
-            )
-
-        chosen_method = "fft" if times["fft"] < times["direct"] else "direct"
-        return chosen_method, times
-
-    # fftconvolve doesn't support complex256
-    fftconv_unsup = "complex256" if sys.maxsize > 2 ** 32 else "complex192"
-    if hasattr(cp, fftconv_unsup):
-        if volume.dtype == fftconv_unsup or kernel.dtype == fftconv_unsup:
-            return "direct"
-
-    # for integer input,
-    # catch when more precision required than float provides (representing an
-    # integer as float can lose precision in fftconvolve if larger than 2**52)
-    if any([_numeric_arrays([x], kinds="ui") for x in [volume, kernel]]):
-        max_value = int(cp.abs(volume).max()) * int(cp.abs(kernel).max())
-        max_value *= int(min(volume.size, kernel.size))
-        if max_value > 2 ** cp.finfo("float").nmant - 1:
-            return "direct"
-
-    if _numeric_arrays([volume, kernel], kinds="b"):
-        return "direct"
-
-    if _numeric_arrays([volume, kernel]):
-        if _fftconv_faster(volume, kernel, mode):
-            return "fft"
-
-    return "direct"
-
-
-def convolve(in1, in2, mode="full", method="auto"):
-    """
-    Convolve two N-dimensional arrays.
-
-    Convolve `in1` and `in2`, with the output size determined by the
-    `mode` argument.
-
-    Parameters
-    ----------
-    in1 : array_like
-        First input.
-    in2 : array_like
-        Second input. Should have the same number of dimensions as `in1`.
-    mode : str {'full', 'valid', 'same'}, optional
-        A string indicating the size of the output:
-
-        ``full``
-           The output is the full discrete linear convolution
-           of the inputs. (Default)
-        ``valid``
-           The output consists only of those elements that do not
-           rely on the zero-padding. In 'valid' mode, either `in1` or `in2`
-           must be at least as large as the other in every dimension.
-        ``same``
-           The output is the same size as `in1`, centered
-           with respect to the 'full' output.
-    method : str {'auto', 'direct', 'fft'}, optional
-        A string indicating which method to use to calculate the convolution.
-
-        ``direct``
-           The convolution is determined directly from sums, the definition of
-           convolution.
-        ``fft``
-           The Fourier Transform is used to perform the convolution by calling
-           `fftconvolve`.
-        ``auto``
-           Automatically chooses direct or Fourier method based on an estimate
-           of which is faster (default).  See Notes for more detail.
-
-           .. versionadded:: 0.19.0
-
-    Returns
-    -------
-    convolve : array
-        An N-dimensional array containing a subset of the discrete linear
-        convolution of `in1` with `in2`.
-
-    See Also
-    --------
-    numpy.polymul : performs polynomial multiplication (same operation, but
-                    also accepts poly1d objects)
-    choose_conv_method : chooses the fastest appropriate convolution method
-    fftconvolve
-
-    Notes
-    -----
-    By default, `convolve` and `correlate` use ``method='auto'``, which calls
-    `choose_conv_method` to choose the fastest method using pre-computed
-    values (`choose_conv_method` can also measure real-world timing with a
-    keyword argument). Because `fftconvolve` relies on floating point numbers,
-    there are certain constraints that may force `method=direct` (more detail
-    in `choose_conv_method` docstring).
-
-    Examples
-    --------
-    Smooth a square pulse using a Hann window:
-
-    >>> from scipy import signal
-    >>> sig = np.repeat([0., 1., 0.], 100)
-    >>> win = signal.hann(50)
-    >>> filtered = signal.convolve(sig, win, mode='same') / sum(win)
-
-    >>> import matplotlib.pyplot as plt
-    >>> fig, (ax_orig, ax_win, ax_filt) = plt.subplots(3, 1, sharex=True)
-    >>> ax_orig.plot(sig)
-    >>> ax_orig.set_title('Original pulse')
-    >>> ax_orig.margins(0, 0.1)
-    >>> ax_win.plot(win)
-    >>> ax_win.set_title('Filter impulse response')
-    >>> ax_win.margins(0, 0.1)
-    >>> ax_filt.plot(filtered)
-    >>> ax_filt.set_title('Filtered signal')
-    >>> ax_filt.margins(0, 0.1)
-    >>> fig.tight_layout()
-    >>> fig.show()
-
-    """
-    volume = asarray(in1)
-    kernel = asarray(in2)
-
-    if volume.ndim == kernel.ndim == 0:
-        return volume * kernel
-    elif volume.ndim != kernel.ndim:
-        raise ValueError("in1 and in2 should have the same dimensionality")
-
-    if _inputs_swap_needed(mode, volume.shape, kernel.shape):
-        # Convolution is commutative; order doesn't have any effect on output
-        volume, kernel = kernel, volume
-
-    if method == "auto":
-        method = choose_conv_method(volume, kernel, mode=mode)
-
-    if method == "fft":
-        out = fftconvolve(volume, kernel, mode=mode)
-        result_type = cp.result_type(volume, kernel)
-        if result_type.kind in {"u", "i"}:
-            out = cp.around(out)
-        return out.astype(result_type)
-    elif method == "direct":
-
-        if volume.ndim > 1:
-            raise ValueError("Direct method is only implemented for 1D")
-
-        swapped_inputs = (mode != "valid") and (kernel.size > volume.size)
-
-        if swapped_inputs:
-            volume, kernel = kernel, volume
-
-        return _signaltools._convolve(
-            volume, kernel, True, swapped_inputs, mode
-        )
-
-    else:
-        raise ValueError(
-            "Acceptable method flags are 'auto'," " 'direct', or 'fft'."
-        )
-
-
-def wiener(im, mysize=None, noise=None):
-    """
-    Perform a Wiener filter on an N-dimensional array.
-
-    Apply a Wiener filter to the N-dimensional array `im`.
-
-    Parameters
-    ----------
-    im : ndarray
-        An N-dimensional array.
-    mysize : int or array_like, optional
-        A scalar or an N-length list giving the size of the Wiener filter
-        window in each dimension.  Elements of mysize should be odd.
-        If mysize is a scalar, then this scalar is used as the size
-        in each dimension.
-    noise : float, optional
-        The noise-power to use. If None, then noise is estimated as the
-        average of the local variance of the input.
-
-    Returns
-    -------
-    out : ndarray
-        Wiener filtered result with the same shape as `im`.
-
-    """
-    im = asarray(im)
-    if mysize is None:
-        mysize = [3] * im.ndim
-    mysize = np.asarray(mysize)
-    if mysize.shape == ():
-        mysize = cp.repeat(mysize.item(), im.ndim)
-        mysize = np.asarray(mysize)
-
-    # Estimate the local mean
-    lMean = correlate(im, ones(mysize), "same") / prod(mysize, axis=0)
-
-    # Estimate the local variance
-    lVar = (
-        correlate(im ** 2, ones(mysize), "same") / prod(mysize, axis=0)
-        - lMean ** 2
-    )
-
-    # Estimate the noise power if needed.
-    if noise is None:
-        noise = mean(ravel(lVar), axis=0)
-
-    res = im - lMean
-    res *= 1 - noise / lVar
-    res += lMean
-    out = where(lVar < noise, lMean, res)
-
-    return out
-
-
-def convolve2d(
-    in1,
-    in2,
-    mode="full",
-    boundary="fill",
-    fillvalue=0,
-    cp_stream=cp.cuda.stream.Stream(null=True),
-    use_numba=False,
-):
-    """
-    Convolve two 2-dimensional arrays.
-    Convolve `in1` and `in2` with output size determined by `mode`, and
-    boundary conditions determined by `boundary` and `fillvalue`.
-    Parameters
-    ----------
-    in1 : array_like
-        First input.
-    in2 : array_like
-        Second input. Should have the same number of dimensions as `in1`.
-    mode : str {'full', 'valid', 'same'}, optional
-        A string indicating the size of the output:
-        ``full``
-           The output is the full discrete linear convolution
-           of the inputs. (Default)
-        ``valid``
-           The output consists only of those elements that do not
-           rely on the zero-padding. In 'valid' mode, either `in1` or `in2`
-           must be at least as large as the other in every dimension.
-        ``same``
-           The output is the same size as `in1`, centered
-           with respect to the 'full' output.
-    boundary : str {'fill', 'wrap', 'symm'}, optional
-        A flag indicating how to handle boundaries:
-        ``fill``
-           pad input arrays with fillvalue. (default)
-        ``wrap``
-           circular boundary conditions.
-        ``symm``
-           symmetrical boundary conditions.
-    fillvalue : scalar, optional
-        Value to fill pad input arrays with. Default is 0.
-    Returns
-    -------
-    out : ndarray
-        A 2-dimensional array containing a subset of the discrete linear
-        convolution of `in1` with `in2`.
-    Examples
-    --------
-    Compute the gradient of an image by 2D convolution with a complex Scharr
-    operator.  (Horizontal operator is real, vertical is imaginary.)  Use
-    symmetric boundary condition to avoid creating edges at the image
-    boundaries.
-    >>> from scipy import signal
-    >>> from scipy import misc
-    >>> ascent = misc.ascent()
-    >>> scharr = np.array([[ -3-3j, 0-10j,  +3 -3j],
-    ...                    [-10+0j, 0+ 0j, +10 +0j],
-    ...                    [ -3+3j, 0+10j,  +3 +3j]]) # Gx + j*Gy
-    >>> grad = signal.convolve2d(ascent, scharr, boundary='symm', mode='same')
-    >>> import matplotlib.pyplot as plt
-    >>> fig, (ax_orig, ax_mag, ax_ang) = plt.subplots(3, 1, figsize=(6, 15))
-    >>> ax_orig.imshow(ascent, cmap='gray')
-    >>> ax_orig.set_title('Original')
-    >>> ax_orig.set_axis_off()
-    >>> ax_mag.imshow(np.absolute(grad), cmap='gray')
-    >>> ax_mag.set_title('Gradient magnitude')
-    >>> ax_mag.set_axis_off()
-    >>> ax_ang.imshow(np.angle(grad), cmap='hsv') # hsv is cyclic, like angles
-    >>> ax_ang.set_title('Gradient orientation')
-    >>> ax_ang.set_axis_off()
-    >>> fig.show()
-    """
-    in1 = asarray(in1)
-    in2 = asarray(in2)
-
-    if not in1.ndim == in2.ndim == 2:
-        raise ValueError("convolve2d inputs must both be 2D arrays")
-
-    if _inputs_swap_needed(mode, in1.shape, in2.shape):
-        in1, in2 = in2, in1
-
-    out = _signaltools._convolve2d(
-        in1,
-        in2,
-        1,
-        mode,
-        boundary,
-        fillvalue,
-        cp_stream=cp_stream,
-        use_numba=use_numba,
-    )
-    return out
-
-
-def correlate2d(
-    in1,
-    in2,
-    mode="full",
-    boundary="fill",
-    fillvalue=0,
-    cp_stream=cp.cuda.stream.Stream(null=True),
-    use_numba=False,
-):
-    """
-    Cross-correlate two 2-dimensional arrays.
-    Cross correlate `in1` and `in2` with output size determined by `mode`, and
-    boundary conditions determined by `boundary` and `fillvalue`.
-    Parameters
-    ----------
-    in1 : array_like
-        First input.
-    in2 : array_like
-        Second input. Should have the same number of dimensions as `in1`.
-    mode : str {'full', 'valid', 'same'}, optional
-        A string indicating the size of the output:
-        ``full``
-           The output is the full discrete linear cross-correlation
-           of the inputs. (Default)
-        ``valid``
-           The output consists only of those elements that do not
-           rely on the zero-padding. In 'valid' mode, either `in1` or `in2`
-           must be at least as large as the other in every dimension.
-        ``same``
-           The output is the same size as `in1`, centered
-           with respect to the 'full' output.
-    boundary : str {'fill', 'wrap', 'symm'}, optional
-        A flag indicating how to handle boundaries:
-        ``fill``
-           pad input arrays with fillvalue. (default)
-        ``wrap``
-           circular boundary conditions.
-        ``symm``
-           symmetrical boundary conditions.
-    fillvalue : scalar, optional
-        Value to fill pad input arrays with. Default is 0.
-    Returns
-    -------
-    correlate2d : ndarray
-        A 2-dimensional array containing a subset of the discrete linear
-        cross-correlation of `in1` with `in2`.
-    Examples
-    --------
-    Use 2D cross-correlation to find the location of a template in a noisy
-    image:
-    >>> from scipy import signal
-    >>> from scipy import misc
-    >>> face = misc.face(gray=True) - misc.face(gray=True).mean()
-    >>> template = np.copy(face[300:365, 670:750])  # right eye
-    >>> template -= template.mean()
-    >>> face = face + np.random.randn(*face.shape) * 50  # add noise
-    >>> corr = signal.correlate2d(face, template, boundary='symm', mode='same')
-    >>> y, x = np.unravel_index(np.argmax(corr), corr.shape)  # find the match
-    >>> import matplotlib.pyplot as plt
-    >>> fig, (ax_orig, ax_template, ax_corr) = plt.subplots(3, 1,
-    ...                                                     figsize=(6, 15))
-    >>> ax_orig.imshow(face, cmap='gray')
-    >>> ax_orig.set_title('Original')
-    >>> ax_orig.set_axis_off()
-    >>> ax_template.imshow(template, cmap='gray')
-    >>> ax_template.set_title('Template')
-    >>> ax_template.set_axis_off()
-    >>> ax_corr.imshow(corr, cmap='gray')
-    >>> ax_corr.set_title('Cross-correlation')
-    >>> ax_corr.set_axis_off()
-    >>> ax_orig.plot(x, y, 'ro')
-    >>> fig.show()
-    """
-    in1 = asarray(in1)
-    in2 = asarray(in2)
-
-    if not in1.ndim == in2.ndim == 2:
-        raise ValueError("correlate2d inputs must both be 2D arrays")
-
-    swapped_inputs = _inputs_swap_needed(mode, in1.shape, in2.shape)
-    if swapped_inputs:
-        in1, in2 = in2, in1
-
-    out = _signaltools._convolve2d(
-        in1,
-        in2.conj(),
-        0,
-        mode,
-        boundary,
-        fillvalue,
-        cp_stream=cp_stream,
-        use_numba=use_numba,
-    )
-
-    if swapped_inputs:
-        out = out[::-1, ::-1]
-
-    return out
-
-
-def lfiltic(b, a, y, x=None):
-    """
-    Construct initial conditions for lfilter given input and output vectors.
-
-    Given a linear filter (b, a) and initial conditions on the output `y`
-    and the input `x`, return the initial conditions on the state vector zi
-    which is used by `lfilter` to generate the output given the input.
-
-    Parameters
-    ----------
-    b : array_like
-        Linear filter term.
-    a : array_like
-        Linear filter term.
-    y : array_like
-        Initial conditions.
-
-        If ``N = len(a) - 1``, then ``y = {y[-1], y[-2], ..., y[-N]}``.
-
-        If `y` is too short, it is padded with zeros.
-    x : array_like, optional
-        Initial conditions.
-
-        If ``M = len(b) - 1``, then ``x = {x[-1], x[-2], ..., x[-M]}``.
-
-        If `x` is not given, its initial conditions are assumed zero.
-
-        If `x` is too short, it is padded with zeros.
-
-    Returns
-    -------
-    zi : ndarray
-        The state vector ``zi = {z_0[-1], z_1[-1], ..., z_K-1[-1]}``,
-        where ``K = max(M, N)``.
-
-    See Also
-    --------
-    lfilter, lfilter_zi
-
-    """
-    N = cp.size(a) - 1
-    M = cp.size(b) - 1
-    K = cp.max(M, N)
-    y = asarray(y)
-    if y.dtype.kind in "bui":
-        # ensure calculations are floating point
-        y = y.astype(np.float64)
-    zi = zeros(K, y.dtype)
-    if x is None:
-        x = zeros(M, y.dtype)
-    else:
-        x = asarray(x)
-        L = cp.size(x)
-        if L < M:
-            x = r_[x, zeros(M - L)]
-    L = cp.size(y)
-    if L < N:
-        y = r_[y, zeros(N - L)]
-
-    for m in range(M):
-        zi[m] = cp.sum(b[m + 1 :] * x[: M - m], axis=0)
-
-    for m in range(N):
-        zi[m] -= cp.sum(a[m + 1 :] * y[: N - m], axis=0)
-
-    return zi
-
-
-def hilbert(x, N=None, axis=-1):
-    """
-    Compute the analytic signal, using the Hilbert transform.
-
-    The transformation is done along the last axis by default.
-
-    Parameters
-    ----------
-    x : array_like
-        Signal data.  Must be real.
-    N : int, optional
-        Number of Fourier components.  Default: ``x.shape[axis]``
-    axis : int, optional
-        Axis along which to do the transformation.  Default: -1.
-
-    Returns
-    -------
-    xa : ndarray
-        Analytic signal of `x`, of each 1-D array along `axis`
-
-    See Also
-    --------
-    scipy.fftpack.hilbert : Return Hilbert transform of a periodic sequence x.
-
-    Notes
-    -----
-    The analytic signal ``x_a(t)`` of signal ``x(t)`` is:
-
-    .. math:: x_a = F^{-1}(F(x) 2U) = x + i y
-
-    where `F` is the Fourier transform, `U` the unit step function,
-    and `y` the Hilbert transform of `x`. [1]_
-
-    In other words, the negative half of the frequency spectrum is zeroed
-    out, turning the real-valued signal into a complex signal.  The Hilbert
-    transformed signal can be obtained from ``np.imag(hilbert(x))``, and the
-    original signal from ``np.real(hilbert(x))``.
-
-    Examples
-    ---------
-    In this example we use the Hilbert transform to determine the amplitude
-    envelope and instantaneous frequency of an amplitude-modulated signal.
-
-    >>> import numpy as np
-    >>> import matplotlib.pyplot as plt
-    >>> from scipy.signal import hilbert, chirp
-
-    >>> duration = 1.0
-    >>> fs = 400.0
-    >>> samples = int(fs*duration)
-    >>> t = np.arange(samples) / fs
-
-    We create a chirp of which the frequency increases from 20 Hz to 100 Hz and
-    apply an amplitude modulation.
-
-    >>> signal = chirp(t, 20.0, t[-1], 100.0)
-    >>> signal *= (1.0 + 0.5 * np.sin(2.0*np.pi*3.0*t) )
-
-    The amplitude envelope is given by magnitude of the analytic signal. The
-    instantaneous frequency can be obtained by differentiating the
-    instantaneous phase in respect to time. The instantaneous phase corresponds
-    to the phase angle of the analytic signal.
-
-    >>> analytic_signal = hilbert(signal)
-    >>> amplitude_envelope = np.abs(analytic_signal)
-    >>> instantaneous_phase = np.unwrap(np.angle(analytic_signal))
-    >>> instantaneous_frequency = (np.diff(instantaneous_phase) /
-    ...                            (2.0*np.pi) * fs)
-
-    >>> fig = plt.figure()
-    >>> ax0 = fig.add_subplot(211)
-    >>> ax0.plot(t, signal, label='signal')
-    >>> ax0.plot(t, amplitude_envelope, label='envelope')
-    >>> ax0.set_xlabel("time in seconds")
-    >>> ax0.legend()
-    >>> ax1 = fig.add_subplot(212)
-    >>> ax1.plot(t[1:], instantaneous_frequency)
-    >>> ax1.set_xlabel("time in seconds")
-    >>> ax1.set_ylim(0.0, 120.0)
-
-    References
-    ----------
-    .. [1] Wikipedia, "Analytic signal".
-           https://en.wikipedia.org/wiki/Analytic_signal
-    .. [2] Leon Cohen, "Time-Frequency Analysis", 1995. Chapter 2.
-    .. [3] Alan V. Oppenheim, Ronald W. Schafer. Discrete-Time Signal
-           Processing, Third Edition, 2009. Chapter 12.
-           ISBN 13: 978-1292-02572-8
-
-    """
-    x = asarray(x)
-    if iscomplexobj(x):
-        raise ValueError("x must be real.")
-    if N is None:
-        N = x.shape[axis]
-    if N <= 0:
-        raise ValueError("N must be positive.")
-
-    Xf = fftpack.fft(x, N, axis=axis)
-    h = zeros(N)
-    if N % 2 == 0:
-        h[0] = h[N // 2] = 1
-        h[1 : N // 2] = 2
-    else:
-        h[0] = 1
-        h[1 : (N + 1) // 2] = 2
-
-    if x.ndim > 1:
-        ind = [newaxis] * x.ndim
-        ind[axis] = slice(None)
-        h = h[tuple(ind)]
-    x = fftpack.ifft(Xf * h, axis=axis)
-    return x
-
-
-def hilbert2(x, N=None):
-    """
-    Compute the '2-D' analytic signal of `x`
-
-    Parameters
-    ----------
-    x : array_like
-        2-D signal data.
-    N : int or tuple of two ints, optional
-        Number of Fourier components. Default is ``x.shape``
-
-    Returns
-    -------
-    xa : ndarray
-        Analytic signal of `x` taken along axes (0,1).
-
-    References
-    ----------
-    .. [1] Wikipedia, "Analytic signal",
-        https://en.wikipedia.org/wiki/Analytic_signal
-
-    """
-    x = atleast_2d(x)
-    if x.ndim > 2:
-        raise ValueError("x must be 2-D.")
-    if iscomplexobj(x):
-        raise ValueError("x must be real.")
-    if N is None:
-        N = x.shape
-    elif isinstance(N, int):
-        if N <= 0:
-            raise ValueError("N must be positive.")
-        N = (N, N)
-    elif len(N) != 2 or cp.any(cp.asarray(N) <= 0):
-        raise ValueError(
-            "When given as a tuple, N must hold exactly two positive integers"
-        )
-
-    Xf = fftpack.fft2(x, N, axes=(0, 1))
-    h1 = zeros(N[0], "d")
-    h2 = zeros(N[1], "d")
-    for p in range(2):
-        h = eval("h%d" % (p + 1))
-        N1 = N[p]
-        if N1 % 2 == 0:
-            h[0] = h[N1 // 2] = 1
-            h[1 : N1 // 2] = 2
-        else:
-            h[0] = 1
-            h[1 : (N1 + 1) // 2] = 2
-        exec("h%d = h" % (p + 1), globals(), locals())
-
-    h = h1[:, newaxis] * h2[newaxis, :]
-    k = x.ndim
-    while k > 2:
-        h = h[:, newaxis]
-        k -= 1
-    x = fftpack.ifft2(Xf * h, axes=(0, 1))
-    return x
 
 
 def cmplx_sort(p):
@@ -1297,127 +40,9 @@ def cmplx_sort(p):
     array([0, 2, 3, 1])
 
     """
-    p = asarray(p)
-    if iscomplexobj(p):
-        indx = argsort(abs(p))
+    p = cp.asarray(p)
+    if cp.iscomplexobj(p):
+        indx = cp.argsort(abs(p))
     else:
-        indx = argsort(p)
-    return take(p, indx, 0), indx
-
-
-def detrend(data, axis=-1, type="linear", bp=0, overwrite_data=False):
-    """
-    Remove linear trend along axis from data.
-
-    Parameters
-    ----------
-    data : array_like
-        The input data.
-    axis : int, optional
-        The axis along which to detrend the data. By default this is the
-        last axis (-1).
-    type : {'linear', 'constant'}, optional
-        The type of detrending. If ``type == 'linear'`` (default),
-        the result of a linear least-squares fit to `data` is subtracted
-        from `data`.
-        If ``type == 'constant'``, only the mean of `data` is subtracted.
-    bp : array_like of ints, optional
-        A sequence of break points. If given, an individual linear fit is
-        performed for each part of `data` between two break points.
-        Break points are specified as indices into `data`.
-    overwrite_data : bool, optional
-        If True, perform in place detrending and avoid a copy. Default is False
-
-    Returns
-    -------
-    ret : ndarray
-        The detrended input data.
-
-    Examples
-    --------
-    >>> from scipy import signal
-    >>> randgen = np.random.RandomState(9)
-    >>> npoints = 1000
-    >>> noise = randgen.randn(npoints)
-    >>> x = 3 + 2*np.linspace(0, 1, npoints) + noise
-    >>> (signal.detrend(x) - noise).max() < 0.01
-    True
-
-    """
-    if type not in ["linear", "l", "constant", "c"]:
-        raise ValueError("Trend type must be 'linear' or 'constant'.")
-    data = asarray(data)
-    dtype = data.dtype.char
-    if dtype not in "dfDF":
-        dtype = "d"
-    if type in ["constant", "c"]:
-        ret = data - expand_dims(mean(data, axis), axis)
-        return ret
-    else:
-        dshape = data.shape
-        N = dshape[axis]
-        bp = sort(unique(r_[0, bp, N]))
-        if cp.any(bp > N):
-            raise ValueError(
-                "Breakpoints must be less than length of \
-                data along given axis."
-            )
-        Nreg = len(bp) - 1
-        # Restructure data so that axis is along first dimension and
-        #  all other dimensions are collapsed into second dimension
-        rnk = len(dshape)
-        if axis < 0:
-            axis = axis + rnk
-        newdims = np.r_[axis, 0:axis, axis + 1 : rnk]
-        newdata = reshape(
-            transpose(data, tuple(newdims)), (N, _prod(dshape) // N)
-        )
-        if not overwrite_data:
-            newdata = newdata.copy()  # make sure we have a copy
-        if newdata.dtype.char not in "dfDF":
-            newdata = newdata.astype(dtype)
-        # Find leastsq fit and remove it for each piece
-        for m in range(Nreg):
-            Npts = int(bp[m + 1] - bp[m])
-            A = ones((Npts, 2), dtype)
-            A[:, 0] = arange(1, Npts + 1) * 1.0 / Npts
-            sl = slice(bp[m], bp[m + 1])
-            coef, resids, rank, s = linalg.lstsq(A, newdata[sl])
-            newdata[sl] = newdata[sl] - dot(A, coef)
-        # Put data back in original shape.
-        tdshape = take(asarray(dshape), asarray(newdims), 0)
-        ret = reshape(newdata, tuple(cp.asnumpy(tdshape)))
-        vals = list(range(1, rnk))
-        olddims = vals[:axis] + [0] + vals[axis:]
-        ret = transpose(ret, tuple(cp.asnumpy(olddims)))
-        return ret
-
-
-def freq_shift(x, freq, fs):
-    """
-    Frequency shift signal by freq at fs sample rate
-
-    Parameters
-    ----------
-    x : array_like, complex valued
-        The data to be shifted.
-    freq : float
-        Shift by this many (Hz)
-    fs : float
-        Sampling rate of the signal
-    domain : string
-        freq or time
-    """
-    x = asarray(x)
-    return x * exp(-1j * 2 * pi * freq / fs * arange(x.size))
-
-
-def _prod(iterable):
-    """
-    Product of a list of numbers.
-    Faster than cp.prod for short lists like array shapes.
-    """
-    product = 1
-    for x in iterable:
-        product *= x
-    return product
+        indx = cp.argsort(p)
+    return cp.take(p, indx, 0), indx
